feat(service) business validations and exceptions
- added custom exceptions
- business validations in same methods
- now te update method of all services modify
the entitie data with the update constructor

+++ src/main/java/com/app/petclinic_scheduler/service/CustomerService.java
@@ -3,6 +3,8 @@ package com.app.petclinic_scheduler.service;
 import com.app.petclinic_scheduler.dto.customer.CustomerRequestDTO;
 import com.app.petclinic_scheduler.dto.customer.CustomerResponseDTO;
 import com.app.petclinic_scheduler.dto.customer.CustomerSummaryDTO;
+import com.app.petclinic_scheduler.exception.BussinesException;
+import com.app.petclinic_scheduler.exception.ResourceNotFoundException;
 import com.app.petclinic_scheduler.model.Customer;
 import com.app.petclinic_scheduler.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,12 +34,20 @@ public class CustomerService {
     }
 
     public Optional<CustomerResponseDTO> findByCpf(String cpf) {
+        if (!customerRepository.existsByCpf(cpf)) {
+            throw new BussinesException("Cliente com cpf: " + cpf + " não encontrado");
+        }
+
         return customerRepository
                 .findByCpf(cpf)
                 .map(CustomerResponseDTO::new);
     }
 
     public void saveCustomer(CustomerRequestDTO customerData) {
+        if (customerRepository.existsByCpf(customerData.cpf())) {
+            throw new BussinesException("Já existe um cliente cadastrado com o cpf: " + customerData.cpf());
+        }
+
         customerRepository
                 .save(new Customer(customerData));
     }
@@ -45,24 +55,19 @@ public class CustomerService {
     public void updateCustomer (UUID id, CustomerRequestDTO updatedData) {
         Customer existingCustomer = customerRepository
                 .findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new ResourceNotFoundException("Cliente com id: " + id + " não encontrado"));
 
-        existingCustomer
-                .setCpf(updatedData.cpf());
-        existingCustomer
-                .setName(updatedData.name());
-        existingCustomer
-                .setEmail(updatedData.email());
-        existingCustomer
-                .setAddress(updatedData.address());
-        existingCustomer
-                .setPhoneNumber(updatedData.phoneNumber());
+        existingCustomer.UpdateFromDTO(updatedData);
 
         customerRepository
                 .save(existingCustomer);
     }
 
     public void deleteCustomer(UUID id) {
+        if(!customerRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Cliente com id: " + id + " não encontrado");
+        }
+
         customerRepository
                 .deleteById(id);
     }
