+++ src/main/java/com/app/petclinic_scheduler/service/ServiceProvideService.java
@@ -3,6 +3,8 @@ package com.app.petclinic_scheduler.service;
 import com.app.petclinic_scheduler.dto.provided_services.ProvidedServiceRequestDTO;
 import com.app.petclinic_scheduler.dto.provided_services.ProvidedServiceResponseDTO;
 import com.app.petclinic_scheduler.dto.provided_services.ProvidedServiceSummaryDTO;
+import com.app.petclinic_scheduler.exception.BussinesException;
+import com.app.petclinic_scheduler.exception.ResourceNotFoundException;
 import com.app.petclinic_scheduler.model.ProvidedService;
 import com.app.petclinic_scheduler.model.Scheduling;
 import com.app.petclinic_scheduler.repository.ProvidedServiceRepository;
@@ -32,35 +34,39 @@ public class ServiceProvideService {
     }
 
     public Optional<ProvidedServiceResponseDTO> findById(UUID id) {
+        if(!providedServiceRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Serviço com o id: " + id + " não encontrado");
+        }
+
         return providedServiceRepository
                 .findById(id)
                 .map(ProvidedServiceResponseDTO::new);
     }
 
     public void saveProvidedService(ProvidedServiceRequestDTO provideServicesData) {
-        Optional<Scheduling> scheduling = schedulingRepository
-                .findById(provideServicesData
-                        .schedulingId());
+        UUID schedulingId = provideServicesData.schedulingId();
+
+        Scheduling scheduling = schedulingRepository
+                .findById(schedulingId)
+                .orElseThrow(() -> new BussinesException("Agendamento com o id: " + schedulingId + " não encontrado" ));
 
         providedServiceRepository
-                .save(new ProvidedService(provideServicesData, scheduling
-                        .orElseThrow()));
+                .save(new ProvidedService(provideServicesData, scheduling));
     }
 
     public void updateProvidedService(UUID id, ProvidedServiceRequestDTO updateData) {
         ProvidedService existingProvideService = providedServiceRepository
                 .findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new ResourceNotFoundException("Serviço com o id: " + id + " não encontrado"));
 
-        existingProvideService
-                .setDescription(updateData.description());
-        existingProvideService
-                .setPrice(updateData.price());
-        existingProvideService
-                .setDurationInMinutes(updateData.durationInMinutes());
+        existingProvideService.updateFromDTO(updateData);
     }
 
     public void deleteProvidedService(UUID id) {
+        if(!providedServiceRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Serviço com o id: " + id + " não encontrado");
+        }
+
         providedServiceRepository.deleteById(id);
     }
 }
